requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
boto3==1.34.122
python-dotenv==1.0.1
pydantic==2.10.4
requests==2.31.0
streamlit==1.33.0
pypdf==4.2.0
faiss-cpu==1.13.2
psycopg[binary]==3.2.3
